requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.1.0
opencv-python>=4.8.0
boto3>=1.29.0
chromadb>=0.5.0
requests>=2.31.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
